codes/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.13.0=py38haa95532_0
aiohttp=3.7.4=py38h294d835_0
astunparse=1.6.3=pypi_0
async-timeout=3.0.1=py_1000
attrs=21.2.0=pyhd8ed1ab_0
blas=1.0=mkl
blas-devel=3.9.0=9_mkl
blinker=1.4=py38haa95532_0
blosc=1.21.0=h19a0ad4_0
boto=2.49.0=py_0
boto3=1.17.88=pyhd8ed1ab_0
botocore=1.20.88=pyhd8ed1ab_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h294d835_1001
bz2file=0.98=py_0
bzip2=1.0.8=he774522_0
ca-certificates=2021.7.5=haa95532_1
cachetools=4.2.2=pyhd8ed1ab_0
catalogue=2.0.4=py38haa244fe_0
certifi=2021.5.30=py38haa95532_0
cffi=1.14.5=py38hd8c33c5_0
cfitsio=3.470=he774522_6
chardet=4.0.0=py38haa244fe_1
charls=2.2.0=h6c2663c_0
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=3.4.7=py38hd7da0ea_0
cudatoolkit=11.1.1=heb2d755_7
cupy=9.1.0=py38h5fb355a_0
cycler=0.10.0=py38_0
cymem=2.0.5=py38h885f38d_1
cython-blis=0.7.4=py38h347fdf6_0
cytoolz=0.11.0=py38he774522_0
dask-core=2021.6.2=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_1
decorator=4.4.2=pyhd3eb1b0_0
deepface=0.0.65=pypi_0
en-core-web-sm=3.0.0=pypi_0
fastrlock=0.6=py38h885f38d_0
filelock=3.0.12=pypi_0
flask=2.0.1=pypi_0
flatbuffers=1.12=pypi_0
freetype=2.10.4=h546665d_1
fsspec=2021.6.0=pyhd3eb1b0_0
gast=0.4.0=pypi_0
gdown=3.13.0=pypi_0
giflib=5.2.1=h62dcd97_0
google-api-core=1.26.3=pyhd8ed1ab_0
google-auth=1.30.0=pyh44b312d_0
google-auth-oauthlib=0.4.1=py_2
google-cloud-core=1.5.0=pyhd3deb0d_0
google-cloud-storage=1.19.0=py_0
google-crc32c=1.1.2=py38h554a69a_0
google-pasta=0.2.0=pypi_0
google-resumable-media=1.2.0=pyhd3deb0d_0
googleapis-common-protos=1.53.0=py38h5b57dd5_0
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyh9f0ad1d_0
imagecodecs=2021.6.8=py38h5da4933_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py38haa95532_0
intel-openmp=2021.2.0=h57928b3_616
itsdangerous=2.0.1=pypi_0
jinja2=3.0.1=pyhd8ed1ab_0
jmespath=0.10.0=pyh9f0ad1d_0
jpeg=9b=hb83a4c4_2
keras=2.4.3=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py38hd77b12b_0
lcms2=2.12=h83e58a3_0
lerc=2.2.1=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libblas=3.9.0=9_mkl
libcblas=3.9.0=9_mkl
libcrc32c=1.1.1=h0e60522_2
libdeflate=1.7=h2bbff1b_5
liblapack=3.9.0=9_mkl
liblapacke=3.9.0=9_mkl
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h1d00b33_2
libprotobuf=3.17.2=h7755175_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.41.0=h8ffe710_0
libzopfli=1.0.3=ha925a31_0
locket=0.2.1=py38haa95532_1
lz4-c=1.9.3=h8ffe710_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=py38haa95532_0
markupsafe=2.0.1=py38h294d835_0
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
mkl=2021.2.0=hb70f87d_389
mkl-devel=2021.2.0=h57928b3_390
mkl-include=2021.2.0=hb70f87d_389
mkl-service=2.3.0=py38h2bbff1b_1
msys2-conda-epoch=20160418=1
mtcnn=0.1.1=pypi_0
multidict=5.1.0=py38h294d835_1
murmurhash=1.0.5=py38h885f38d_0
networkx=2.5.1=pyhd3eb1b0_0
ninja=1.10.2=h5362a0b_0
numpy=1.21.1=pypi_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.0.1=py38h2a7c758_0
opencv-python=4.5.3.56=pypi_0
openjpeg=2.3.0=h5ec785f_1
openssl=1.1.1k=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyh44b312d_0
pandas=1.2.5=py38hd77b12b_0
partd=1.2.0=pyhd3eb1b0_0
pathy=0.5.2=pyhd8ed1ab_0
pillow=8.2.0=py38h4fa10fc_0
pip=21.1.1=py38haa95532_0
preshed=3.0.5=py38h885f38d_0
protobuf=3.17.2=py38h885f38d_0
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=pyh9f0ad1d_2
pydantic=1.7.3=py38h294d835_1
pyjwt=2.1.0=py38haa95532_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38ha925a31_4
pysocks=1.7.1=py38haa244fe_3
python=3.8.10=hdbf39b2_7
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.8.1=py3.8_cuda11.1_cudnn8_0
pytz=2021.1=pyhd8ed1ab_0
pywavelets=1.1.1=py38he774522_2
pyyaml=5.4.1=py38h2bbff1b_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3deb0d_0
requests-oauthlib=1.3.0=py_0
retina-face=0.0.4=pypi_0
rsa=4.7.2=pyh44b312d_0
s3transfer=0.4.2=pyhd8ed1ab_0
scikit-image=0.18.1=py38hf11a4ad_0
scipy=1.6.2=py38h66253e8_1
setuptools=52.0.0=py38haa95532_0
shellingham=1.4.0=pyh44b312d_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=pypi_0
smart_open=2.2.1=pyh9f0ad1d_0
snappy=1.1.8=h33f27b4_0
spacy=3.0.6=py38h2f20550_0
spacy-legacy=3.0.5=pyhd8ed1ab_0
sqlite=3.35.4=h2bbff1b_0
srsly=2.4.1=py38h885f38d_0
tbb=2021.2.0=h2d74725_0
tensorboard=2.5.0=py_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
thinc=8.0.3=py38h2f20550_1
tifffile=2021.4.8=pyhd3eb1b0_2
tk=8.6.10=h8ffe710_1
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cu111
tornado=6.1=py38h2bbff1b_0
tqdm=4.61.0=pyhd8ed1ab_0
typer=0.3.2=pyhd8ed1ab_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.5=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=0.8.2=pyh44b312d_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_2
wincertstore=0.2=py38_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yarl=1.6.3=py38h294d835_1
zfp=0.5.5=hd77b12b_6
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_1010
zstd=1.4.9=h6255e5f_0
